Construction Search Field + code cleanup
