+++ UnityProject/Assets/Scripts/Core/Utils/MultiInterestBool.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 45f8bc65dd0845a4ac5049c305c7d5b4
+timeCreated: 1721053668
