+++ UnityProject/Assets/Scripts/UI/Systems/MainHUD/UI Bottom/Construction/BuildMenu.cs
@@ -1,6 +1,7 @@
 using Construction;
 using UnityEngine;
 using Construction.Conveyors;
+using Core.Utils;
 using TMPro;
 
 namespace UI.UI_Bottom
@@ -27,6 +28,33 @@ namespace UI.UI_Bottom
 		[Tooltip("The number of the specified item to make ")]
 		[SerializeField] public TMP_InputField NumberInputField = null;
 
+		[SerializeField] private TMP_InputField searchField;
+
+
+		private void Start()
+		{
+			searchField.onValueChanged.AddListener(Search);
+		}
+
+		private void Search(string newValue)
+		{
+			if (string.IsNullOrEmpty(newValue))
+			{
+				foreach (Transform child in contentPanel.transform)
+				{
+					child.SetActive(true);
+				}
+				return;
+			}
+			foreach (Transform child in contentPanel.transform)
+			{
+				string childName = child.name.ToLower();
+				string searchValue = newValue.ToLower();
+				bool isMatch = childName.Contains(searchValue) || Utils.LevenshitenDistance(childName, searchValue) <= 2;
+				child.gameObject.SetActive(isMatch);
+			}
+		}
+
 
 		//TODO: Implement, model kinda after dev spawner.
 
@@ -39,6 +67,7 @@ namespace UI.UI_Bottom
 			conveyorBuildMenu.gameObject.SetActive(false);
 			transform.GetChild(0).gameObject.SetActive(true);
 			currentBuildingMaterial = buildingMaterial;
+			UIManager.Instance.isInputFocus = true;
 			// delete previous results
 			foreach (Transform child in contentPanel.transform)
 			{
@@ -68,6 +97,7 @@ namespace UI.UI_Bottom
 		{
 			_ = SoundManager.Play(CommonSounds.Instance.Click01);
 			transform.GetChild(0).gameObject.SetActive(false);
+			UIManager.Instance.isInputFocus = false;
 		}
 
 		// add a list item to the content panel for spawning the specified result
@@ -79,6 +109,7 @@ namespace UI.UI_Bottom
 			listItem.GetComponent<BuildMenuEntryController>().Initialize(entry, currentBuildingMaterial);
 			listItem.transform.SetParent(contentPanel.transform);
 			listItem.transform.localScale = Vector3.one;
+			listItem.name = entry.Name;
 		}
 	}
 }
